:construction: Playing around with ebiten

+++ pkg/apu/apu.go
@@ -9,6 +9,6 @@ func (apu *APU) Clock() {
 func (apu *APU) Reset() {
 }
 
-func (apu *APU) GetAudioSample() float64 {
-	return 0.0
+func (apu *APU) GetAudioSample() int32 {
+	return 0
 }
